chore: rename helpers to testHelpers
